Add check for zsh presence before installing p10k

+++ zsh-config/install.sh
@@ -1,3 +1,8 @@
+if [ "$(command -v zsh)" == "" ]; then
+        echo 'Error: could not find zsh in your PATH. Please install it and try this script again.'
+        exit 1
+fi
+
 rm -rf ~/powerlevel10k
 git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ~/powerlevel10k
 cp ./.zshrc  ~/.zshrc
